add minor changes and updates

+++ src/main/java/in/cypher/playerscore/beans/response/PlayerHistoryResponse.java
@@ -7,7 +7,7 @@ public class PlayerHistoryResponse {
 
     @Getter
     @Setter
-    public int playerScore;
+    public float playerScore;
 
     @Getter
     @Setter
